+++ src/pages/creation/entry/editor/ProjectEntryEditor.tsx
@@ -12,6 +12,7 @@ import { ID } from '../../../../constants'
 import { getSignedUploadAPI } from '../../../../hooks'
 import { colors, fonts } from '../../../../styles'
 import { useMobileMode, useNotification } from '../../../../utils'
+import { dataURLToBlob } from '../../../../utils/image'
 
 type Rules = string
 
@@ -165,41 +166,6 @@ const editorModules = (
     },
   },
 })
-
-function dataURLToBlob(dataURL: string) {
-  const BASE64_MARKER = ';base64,'
-
-  if (dataURL.indexOf(BASE64_MARKER) === -1) {
-    const parts = dataURL.split(',')
-    const contentType = parts[0]?.split(':')[1]
-    const cmp = parts[1]
-    if (!cmp) {
-      return
-    }
-
-    const raw = decodeURIComponent(cmp)
-
-    return new Blob([raw], { type: contentType })
-  }
-
-  const parts = dataURL.split(BASE64_MARKER)
-  const contentType = parts[0]?.split(':')[1]
-  const cmp = parts[1]
-  if (!cmp) {
-    return
-  }
-
-  const raw = window.atob(cmp)
-  const rawLength = raw.length
-
-  const uInt8Array = new Uint8Array(rawLength)
-
-  for (let i = 0; i < rawLength; ++i) {
-    uInt8Array[i] = raw.charCodeAt(i)
-  }
-
-  return new Blob([uInt8Array], { type: contentType })
-}
 
 export const ProjectEntryEditor = ({
   name,
